+++ screens/home/parallax-desktop/index.js
@@ -5,9 +5,10 @@ import { constant } from './constant'
 const assetDomain = process.env.config?.baseEndpoint ?? '',
     assetPrefix = process.env.config?.assetPrefix ?? ''
 
-export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
+export default function HomeScreen({ sectionOne, sectionTwo }) {
     if (!sectionOne) return <></>
     const [isWebpSupport, setIsWebpSupport] = useState(true)
+    const [isLandscape, setIsLandscape] = useState(false)
     const [heroHeight, setHeroHeight] = useState(0)
     const [iconHeight, setIconHeight] = useState({
         inpatient: 0,
@@ -54,6 +55,10 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                 width: (35 / 100) * window.innerWidth,
                 height: (21 / 100) * window.innerWidth,
             })
+
+            if (window.innerWidth < 1024 && window.innerHeight < 450) {
+                setIsLandscape(true)
+            }
         }
     }, [])
 
@@ -243,23 +248,25 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
         familyImg = `${sectionOne?.familyImage?.url ?? ''}`
     }
 
-    if (window) {
-        if (window.innerWidth < 1024 && window.innerHeight < 450) {
-            familyImg = `${
-                sectionOne?.familyImageWebp?.formats?.medium?.url ?? ''
-            }`
-            if (!sectionOne.familyImage || !isWebpSupport) {
-                familyImg = `${
-                    sectionOne?.familyImage?.formats?.medium?.url ?? ''
-                }`
-            }
+    // if (window) {
+    //     if (window.innerWidth < 1024 && window.innerHeight < 450) {
+    //         familyImg = `${
+    //             sectionOne?.familyImageWebp?.formats?.medium?.url ?? ''
+    //         }`
+    //         if (!sectionOne.familyImage || !isWebpSupport) {
+    //             familyImg = `${
+    //                 sectionOne?.familyImage?.formats?.medium?.url ?? ''
+    //             }`
+    //         }
+    //     }
+    // }
+
+    if (isLandscape) {
+        familyImg = `${sectionOne?.familyImageWebp?.formats?.medium?.url ?? ''}`
+        if (!sectionOne.familyImage || !isWebpSupport) {
+            familyImg = `${sectionOne?.familyImage?.formats?.medium?.url ?? ''}`
         }
     }
-    // if (isPortrait) {
-    //     heroImg = header.mobileImage
-    //         ? `${assetDomain}${header.mobileImage.url}`
-    //         : ''
-    // }
 
     return (
         <>
@@ -279,7 +286,10 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                     </div>
                 </div>
                 <div className="hero-image" id="hero-image">
-                    <img src={`${assetDomain}${familyImg}`} />
+                    <img
+                        src={`${assetDomain}${familyImg}`}
+                        alt="family-image"
+                    />
                     {/* <div id="hero-icon-all" className="hero-icon-all"> */}
                     <div
                         id="icon-inpatient"
@@ -302,6 +312,7 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                                 src={`${assetDomain}${
                                     sectionOne?.inpatientIcon?.url ?? ''
                                 }`}
+                                alt="inpatient-icon"
                             />
                         </div>
                     </div>
@@ -326,6 +337,7 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                                 src={`${assetDomain}${
                                     sectionOne?.maternityIcon?.url ?? ''
                                 }`}
+                                alt="maternity-icon"
                             />
                         </div>
                     </div>
@@ -350,6 +362,7 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                                 src={`${assetDomain}${
                                     sectionOne?.outpatientIcon?.url ?? ''
                                 }`}
+                                alt="outpatient-icon"
                             />
                         </div>
                     </div>
@@ -375,6 +388,7 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                                 src={`${assetDomain}${
                                     sectionOne?.dentalIcon?.url ?? ''
                                 }`}
+                                alt="dental-icon"
                             />
                         </div>
                     </div>
@@ -396,6 +410,7 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                         src={`${assetDomain}${
                             sectionTwo?.backgroundImage?.url ?? ''
                         }`}
+                        alt="living-room-image"
                     />
 
                     <img
@@ -404,6 +419,7 @@ export default function HomeScreen({ sectionOne, sectionTwo, isPortrait }) {
                         src={`${assetDomain}${
                             sectionTwo?.familyImage?.url ?? ''
                         }`}
+                        alt="family-revolve-image"
                     />
                 </div>
             )}