+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### [v1.1.0]
+
+> 2021-11-16
+
+-   Read SEO title, description, and keywords from strapi for home, product, and download page
+-   Add home page alt on image and set min height for mobile home
+
 ### [v1.0.1]
 
 > 2021-11-15